// Book_Portal_UI/src/app/Models/author-register-request.ts
export class AuthorRegisterRequest{
        username: string | undefined;
        password: string | undefined;
        email: string | undefined;
        auLname: string | undefined;
        auFname: string | undefined;
        phone: string | undefined;
        address: string | undefined;
        city: string | undefined;
        state:string | undefined;
        zip: string | undefined;
       
       
}